src/requirements.txt
# Env & simulation (AI2-THOR 5.0+ required for ProcTHOR / CreateHouse)
ai2thor>=5.0.0
procthor
gymnasium>=0.29.0
prior

# ProcTHOR runtime deps (not all in procthor's setup.py)
attrs
shapely
decorator>=4.0.2,<5.0
moviepy<2
numpy
matplotlib
trimesh
python-sat
python-fcl
canonicaljson
networkx
scipy
pandas

# Optional: global keyboard capture in demo scripts (so keys work when sim window is focused)
pynput

# LLM (Gemini API for Orchestrator + Scene generator)
google-genai

# Load .env for GEMINI_API_KEY / GOOGLE_API_KEY
python-dotenv

# Backend / schemas
pydantic>=2.0.0

# Web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Image processing
pillow>=10.0.0

# Async/concurrency
aiofiles>=23.2.0